refactor(core): Extract webhook request handler to own file
Builds on top of #10296. Move webhook request handler to own class.
